IDN: use correct type for data_size (cid 1604286)
In function dissect_idn_dmx_data() all use of data_size is based on
the 'int' type. Define it as such, rather than uint16_t.


(cherry picked from commit 9c0d6991ba5a2bef9218e2414f8292fba4b52e64)

+++ epan/dissectors/packet-idn.c
@@ -516,7 +516,7 @@ static int dissect_idn_dmx_data(tvbuff_t *tvb, packet_info *pinfo, int offset, p
 	int *count = config->count;
 	int *base = config->base;
 	int base_value;
-	uint16_t data_size = tvb_reported_length_remaining(tvb, offset);
+	int data_size = tvb_reported_length_remaining(tvb, offset);
 	proto_tree *idn_samples_tree = proto_tree_add_subtree(idn_tree, tvb, offset, data_size, ett_data, NULL, "Channels");
 	proto_tree *idn_dmx_subtree;
